Recover some v0.4.0-draft changes

+++ fairseq_ext/data/indexed_dataset.py
@@ -482,7 +482,7 @@ class MMapIndexedDataset(torch.utils.data.Dataset):
         # FIXME: This casting fix may cause problems later
         if np_array.size > 0 and type(np_array[0]) == np.uint16:
             #return torch.from_numpy(np_array.astype(np.int16)).type(torch.ByteTensor)
-            return torch.from_numpy(np_array.astype(np.int16))
+            return torch.from_numpy(np_array.astype(np.int64))
         else:    
             return torch.from_numpy(np_array)
